create ui module
